production-center/core-application/apis/tiktok_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
PyYAML==6.0.2
google-cloud-secret-manager==2.20.1
google-cloud-bigquery==3.25.0
google-auth==2.34.0
requests==2.32.3
loguru==0.7.2
